requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
attrs==24.2.0
blinker==1.8.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.4
